worker/requirements.txt
# ── Celery & broker ────────────────────────────────────────────
celery>=5.4.0
redis[hiredis]>=5.0.7

# ── HTTP client (call FastAPI engine) ──────────────────────────
httpx>=0.27.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.31
asyncpg>=0.29.0

# ── Neo4j ──────────────────────────────────────────────────────
neo4j>=5.22.0

# ── Env & config ───────────────────────────────────────────────
python-dotenv>=1.0.1
